Cleaned up integration folder structure
