+++ src/main/java/com/gyvex/ezafk/integration/PlayerListNameAdapter.java
@@ -3,14 +3,14 @@ package com.gyvex.ezafk.integration;
 import java.util.UUID;
 import org.bukkit.entity.Player;
 
-interface PlayerListNameAdapter {
-    void removeInvalidEntries();
+public interface PlayerListNameAdapter {
+	void removeInvalidEntries();
 
-    String getBaseName(Player player);
+	String getBaseName(Player player);
 
-    boolean apply(Player player, String targetName);
+	boolean apply(Player player, String targetName);
 
-    boolean restore(UUID uuid);
+	boolean restore(UUID uuid);
 
-    void restoreAll();
+	void restoreAll();
 }
